Move isSafe to Standard
