+++ src/main/scala/variant/Standard.scala
@@ -1,6 +1,9 @@
 package chess
 package variant
 
+import bitboard.Bitboard
+import bitboard.Bitboard.*
+
 case object Standard
     extends Variant(
       id = Variant.Id(1),
@@ -26,6 +29,17 @@ case object Standard
             genNonKing(targets) ::: genSafeKing(targets) ::: genCastling ::: enPassantMoves
           else situation.genEvasions(checkers) ::: enPassantMoves
         if situation.sliderBlockers.nonEmpty || enPassantMoves.nonEmpty then
-          candidates.filter(situation.isSafe(king, _, situation.sliderBlockers))
+          candidates.filter(isSafe(situation, king, _, situation.sliderBlockers))
         else candidates
       )
+
+  // Used for filtering candidate moves that would leave put the king in check.
+  private def isSafe(situation: Situation, king: Pos, move: Move, blockers: Bitboard): Boolean =
+    import situation.{ board, color, us, them }
+    if move.enpassant then
+      val newOccupied = (board.occupied ^ move.orig.bb ^ move.dest.withRankOf(move.orig).bb) | move.dest.bb
+      (king.rookAttacks(newOccupied) & them & (board.rooks ^ board.queens)).isEmpty &&
+      (king.bishopAttacks(newOccupied) & them & (board.bishops ^ board.queens)).isEmpty
+    else if !move.castles || !move.promotes then
+      !(us & blockers).contains(move.orig) || Bitboard.aligned(move.orig, move.dest, king)
+    else true
